Add language select viewmodel & enable adv search functionality #12557
